+++ pkgs/jnigen/lib/src/elements/elements.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:json_annotation/json_annotation.dart';
+import 'package:meta/meta.dart';
 
 // Types to describe java API elements
 
@@ -57,9 +58,9 @@ class Classes implements Element<Classes> {
 // option in java.
 
 @JsonSerializable(createToJson: false)
-class ClassDecl extends ClassMember implements Element<ClassDecl> {
+class ClassDecl with ClassMember, Annotated implements Element<ClassDecl> {
   ClassDecl({
-    this.annotations = const [],
+    this.annotations,
     this.javadoc,
     required this.declKind,
     this.modifiers = const {},
@@ -68,6 +69,7 @@ class ClassDecl extends ClassMember implements Element<ClassDecl> {
     this.methods = const [],
     this.fields = const [],
     this.superclass,
+    this.outerClassBinaryName,
     this.interfaces = const [],
     this.hasStaticInit = false,
     this.hasInstanceInit = false,
@@ -79,7 +81,8 @@ class ClassDecl extends ClassMember implements Element<ClassDecl> {
   @override
   final Set<String> modifiers;
 
-  final List<Annotation> annotations;
+  @override
+  List<Annotation>? annotations;
   final KotlinClass? kotlinClass;
   final KotlinPackage? kotlinPackage;
   final JavaDocComment? javadoc;
@@ -95,6 +98,14 @@ class ClassDecl extends ClassMember implements Element<ClassDecl> {
   /// Will default to java.lang.Object if null by [Linker].
   TypeUsage? superclass;
 
+  final String? outerClassBinaryName;
+
+  /// Outer class's [ClassDecl] obtained from [outerClassBinaryName].
+  ///
+  /// Populated by [Linker].
+  @JsonKey(includeFromJson: false)
+  late final ClassDecl? outerClass;
+
   /// Contains enum constant names if class is an enum,
   /// as obtained by `.values()` method in Java.
   final List<String>? values;
@@ -109,12 +120,6 @@ class ClassDecl extends ClassMember implements Element<ClassDecl> {
   @JsonKey(includeFromJson: false)
   late final int superCount;
 
-  /// Parent's [ClassDecl] obtained from [parentName].
-  ///
-  /// Populated by [Linker].
-  @JsonKey(includeFromJson: false)
-  late final ClassDecl? parent;
-
   /// Final name of this class.
   ///
   /// Populated by [Renamer].
@@ -128,11 +133,19 @@ class ClassDecl extends ClassMember implements Element<ClassDecl> {
   @JsonKey(includeFromJson: false)
   late final String typeClassName;
 
-  /// Type parameters including the ones from its ancestors
+  /// Name of the nullable type class.
+  ///
+  /// Populated by [Renamer].
+  @JsonKey(includeFromJson: false)
+  late final String nullableTypeClassName;
+
+  /// Type parameters including the ones from its outer classes.
+  ///
+  /// For `Foo<T>.Bar<U, V>.Baz<W>` it is [T, U, V, W].
   ///
   /// Populated by [Linker].
   @JsonKey(includeFromJson: false)
-  List<TypeParam> allTypeParams = const [];
+  late final List<TypeParam> allTypeParams;
 
   /// The path which this class is generated in.
   ///
@@ -169,15 +182,8 @@ class ClassDecl extends ClassMember implements Element<ClassDecl> {
 
   bool get isObject => superCount == 0;
 
-  // TODO(https://github.com/dart-lang/native/issues/1544): Use a better
-  // heuristic. Class names can have dollar signs without being nested.
   @JsonKey(includeFromJson: false)
-  late final String? parentName = binaryName.contains(r'$')
-      ? binaryName.splitMapJoin(RegExp(r'\$[^$]+$'), onMatch: (_) => '')
-      : null;
-
-  @JsonKey(includeFromJson: false)
-  late final isNested = parentName != null;
+  bool get isNested => outerClassBinaryName != null;
 
   /// Whether the class is actually only a number of top-level Kotlin Functions.
   bool get isTopLevel => kotlinPackage != null;
@@ -274,20 +280,21 @@ class TypeUsage {
         clonedType = ArrayType.fromJson(clonedTypeJson);
         break;
     }
+    clonedType.annotations = type.annotations;
     return TypeUsage(shorthand: shorthand, kind: kind, typeJson: clonedTypeJson)
       ..type = clonedType;
   }
 }
 
-abstract class ReferredType {
-  const ReferredType();
+abstract class ReferredType with Annotated {
+  ReferredType();
   String get name;
 
   R accept<R>(TypeVisitor<R> v);
 }
 
 class PrimitiveType extends ReferredType {
-  static const _primitives = {
+  static final _primitives = {
     'byte': PrimitiveType._(
       name: 'byte',
       signature: 'B',
@@ -362,17 +369,19 @@ class PrimitiveType extends ReferredType {
     ),
   };
 
-  const PrimitiveType._({
+  PrimitiveType._({
     required this.name,
     required this.signature,
     required this.dartType,
     required this.boxedName,
     required this.cType,
     required this.ffiVarArgType,
-  });
+  }) : annotations = null;
 
   @override
   final String name;
+  @override
+  List<Annotation>? annotations;
 
   final String signature;
   final String dartType;
@@ -394,11 +403,14 @@ class PrimitiveType extends ReferredType {
 class DeclaredType extends ReferredType {
   DeclaredType({
     required this.binaryName,
+    this.annotations,
     this.params = const [],
   });
 
   final String binaryName;
   final List<TypeUsage> params;
+  @override
+  List<Annotation>? annotations;
 
   @JsonKey(includeFromJson: false)
   late ClassDecl classDecl;
@@ -421,10 +433,46 @@ class TypeVar extends ReferredType {
   @JsonKey(includeFromJson: false)
   late final TypeParam origin;
 
-  TypeVar({required this.name});
+  TypeVar({
+    required this.name,
+    this.annotations,
+  });
 
   @override
   String name;
+  @override
+  List<Annotation>? annotations;
+
+  @override
+  bool get isNullable {
+    // A type-var is nullable if its origin is nullable.
+    if (origin.isNullable) {
+      return true;
+    }
+    // If origin is non-null, it has to be explicitly set as nullable.
+    if (!origin.isNullable && !hasNullable) {
+      return false;
+    }
+    return super.isNullable;
+  }
+
+  /// Whether this type-variable has a question mark.
+  ///
+  /// This is different from [isNullable], a type-variable that extends
+  /// `JObject?` is nullable, so to get the reference from an object with this
+  /// type, a null check is needed. However type-variables can have an extra
+  /// question mark, meaning that even if the original type extends `JObject`,
+  /// this is nullable.
+  bool get hasQuestionMark {
+    // If the origin has any nullability set, this will only be nullable if it
+    // is explicitly set to be.
+    if (origin.hasNonNull || origin.hasNullable) {
+      return hasNullable;
+    }
+    // Otherwise it is always nullable unless explicitly set to be
+    // non-nullable.
+    return !hasNonNull;
+  }
 
   factory TypeVar.fromJson(Map<String, dynamic> json) =>
       _$TypeVarFromJson(json);
@@ -437,11 +485,23 @@ class TypeVar extends ReferredType {
 
 @JsonSerializable(createToJson: false)
 class Wildcard extends ReferredType {
-  Wildcard({this.extendsBound, this.superBound});
+  Wildcard({
+    this.extendsBound,
+    this.superBound,
+    this.annotations,
+  });
   TypeUsage? extendsBound, superBound;
 
   @override
+  bool get isNullable =>
+      super.isNullable &&
+      // If the extends bound is non-null, this is non-null.
+      !(extendsBound?.type.hasNonNull ?? false);
+
+  @override
   String get name => '?';
+  @override
+  List<Annotation>? annotations;
 
   factory Wildcard.fromJson(Map<String, dynamic> json) =>
       _$WildcardFromJson(json);
@@ -454,11 +514,16 @@ class Wildcard extends ReferredType {
 
 @JsonSerializable(createToJson: false)
 class ArrayType extends ReferredType {
-  ArrayType({required this.type});
-  TypeUsage type;
+  ArrayType({
+    required this.elementType,
+    this.annotations,
+  });
+  TypeUsage elementType;
 
   @override
-  String get name => '[${type.name}';
+  String get name => '[${elementType.name}';
+  @override
+  List<Annotation>? annotations;
 
   factory ArrayType.fromJson(Map<String, dynamic> json) =>
       _$ArrayTypeFromJson(json);
@@ -469,7 +534,58 @@ class ArrayType extends ReferredType {
   }
 }
 
-abstract class ClassMember {
+mixin Annotated {
+  abstract List<Annotation>? annotations;
+
+  static final nullableAnnotations = [
+    // Taken from https://kotlinlang.org/docs/java-interop.html#nullability-annotations
+    'org.jetbrains.annotations.Nullable',
+    'org.jspecify.nullness.Nullable',
+    'com.android.annotations.Nullable',
+    'androidx.annotation.Nullable',
+    'android.support.annotations.Nullable',
+    'edu.umd.cs.findbugs.annotations.Nullable',
+    'org.eclipse.jdt.annotation.Nullable',
+    'lombok.Nullable',
+    'io.reactivex.rxjava3.annotations.Nullable',
+  ];
+  late final bool hasNullable = () {
+    return annotations?.any((annotation) =>
+            nullableAnnotations.contains(annotation.binaryName) ||
+            annotation.binaryName == 'javax.annotation.Nonnull' &&
+                annotation.properties['when'] == 'ALWAYS') ??
+        false;
+  }();
+
+  static final nonNullAnnotations = [
+    // Taken from https://kotlinlang.org/docs/java-interop.html#nullability-annotations
+    'org.jetbrains.annotations.NotNull',
+    'org.jspecify.nullness.NonNull',
+    'com.android.annotations.NonNull',
+    'androidx.annotation.NonNull',
+    'android.support.annotations.NonNull',
+    'edu.umd.cs.findbugs.annotations.NonNull',
+    'org.eclipse.jdt.annotation.NonNull',
+    'lombok.NonNull',
+    'io.reactivex.rxjava3.annotations.NonNull',
+  ];
+  late final hasNonNull = () {
+    return annotations?.any((annotation) =>
+            nonNullAnnotations.contains(annotation.binaryName) ||
+            annotation.binaryName == 'javax.annotation.Nonnull' &&
+                annotation.properties['when'] == 'ALWAYS') ??
+        false; //FIXME
+  }();
+
+  late final bool isNullable = () {
+    if (hasNullable) {
+      return true;
+    }
+    return !hasNonNull;
+  }();
+}
+
+mixin ClassMember {
   String get name;
   ClassDecl get classDecl;
   Set<String> get modifiers;
@@ -485,9 +601,9 @@ abstract class ClassMember {
 }
 
 @JsonSerializable(createToJson: false)
-class Method extends ClassMember implements Element<Method> {
+class Method with ClassMember, Annotated implements Element<Method> {
   Method({
-    this.annotations = const [],
+    this.annotations,
     this.javadoc,
     this.modifiers = const {},
     required this.name,
@@ -501,8 +617,8 @@ class Method extends ClassMember implements Element<Method> {
   final String name;
   @override
   final Set<String> modifiers;
-
-  final List<Annotation> annotations;
+  @override
+  List<Annotation>? annotations;
   final JavaDocComment? javadoc;
   final List<TypeParam> typeParams;
   List<Param> params;
@@ -549,16 +665,20 @@ class Method extends ClassMember implements Element<Method> {
 }
 
 @JsonSerializable(createToJson: false)
-class Param implements Element<Param> {
+class Param with Annotated implements Element<Param> {
   Param({
-    this.annotations = const [],
+    this.annotations,
     this.javadoc,
     required this.name,
     required this.type,
   });
 
-  final List<Annotation> annotations;
+  @override
+  List<Annotation>? annotations;
   final JavaDocComment? javadoc;
+
+  @override
+  bool get isNullable => type.type.isNullable || super.hasNullable;
 
   // Synthetic methods might not have parameter names.
   @JsonKey(defaultValue: 'synthetic')
@@ -583,9 +703,9 @@ class Param implements Element<Param> {
 }
 
 @JsonSerializable(createToJson: false)
-class Field extends ClassMember implements Element<Field> {
+class Field with ClassMember, Annotated implements Element<Field> {
   Field({
-    this.annotations = const [],
+    this.annotations,
     this.javadoc,
     this.modifiers = const {},
     required this.name,
@@ -598,7 +718,8 @@ class Field extends ClassMember implements Element<Field> {
   @override
   final Set<String> modifiers;
 
-  final List<Annotation> annotations;
+  @override
+  List<Annotation>? annotations;
   final JavaDocComment? javadoc;
   final TypeUsage type;
   final Object? defaultValue;
@@ -624,11 +745,23 @@ class Field extends ClassMember implements Element<Field> {
 }
 
 @JsonSerializable(createToJson: false)
-class TypeParam implements Element<TypeParam> {
-  TypeParam({required this.name, this.bounds = const []});
+class TypeParam with Annotated implements Element<TypeParam> {
+  TypeParam({
+    required this.name,
+    this.bounds = const [],
+    this.annotations,
+  });
 
   final String name;
   final List<TypeUsage> bounds;
+
+  @override
+  List<Annotation>? annotations;
+
+  @override
+  bool get hasNonNull =>
+      // A type param with any non-null bound is non-null.
+      super.hasNonNull || bounds.any((bound) => !bound.type.isNullable);
 
   /// Can either be a [ClassDecl] or a [Method].
   ///
@@ -663,15 +796,111 @@ class JavaDocComment implements Element<JavaDocComment> {
   }
 }
 
+@visibleForTesting
+List<TypePathStep> typePathFromString(String? string) {
+  if (string == null) return const [];
+  const innerClass = 46;
+  assert(innerClass == '.'.codeUnitAt(0));
+  const array = 91;
+  assert(array == '['.codeUnitAt(0));
+  const wildcard = 42;
+  assert(wildcard == '*'.codeUnitAt(0));
+  const digit0 = 48;
+  assert(digit0 == '0'.codeUnitAt(0));
+  const digit9 = 57;
+  assert(digit9 == '9'.codeUnitAt(0));
+  const semicolon = 59;
+  assert(semicolon == ';'.codeUnitAt(0));
+  final typePaths = <TypePathStep>[];
+  var number = 0;
+  for (final codeUnit in string.codeUnits) {
+    switch (codeUnit) {
+      case array:
+        typePaths.add(const ToArrayElement());
+      case wildcard:
+        typePaths.add(const ToWildcardBound());
+      case innerClass:
+        typePaths.add(const ToInnerClass());
+      case >= digit0 && <= digit9:
+        number = number * 10 + codeUnit - digit0;
+      case semicolon:
+        typePaths.add(ToTypeParam(number));
+        number = 0;
+      default:
+        throw const FormatException('Invalid type path');
+    }
+  }
+  return typePaths;
+}
+
+sealed class TypePathStep {
+  const TypePathStep();
+}
+
+final class ToArrayElement extends TypePathStep {
+  const ToArrayElement();
+  @override
+  String toString() {
+    return '[';
+  }
+}
+
+final class ToInnerClass extends TypePathStep {
+  const ToInnerClass();
+  @override
+  String toString() {
+    return '.';
+  }
+}
+
+final class ToWildcardBound extends TypePathStep {
+  const ToWildcardBound();
+  @override
+  String toString() {
+    return '*';
+  }
+}
+
+final class ToTypeParam extends TypePathStep {
+  final int index;
+  const ToTypeParam(this.index);
+  @override
+  String toString() {
+    return '$index;';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return other is ToTypeParam && index == other.index;
+  }
+
+  @override
+  int get hashCode => (ToTypeParam).hashCode ^ index.hashCode;
+}
+
 @JsonSerializable(createToJson: false)
 class Annotation implements Element<Annotation> {
-  Annotation({
+  /// Specifies that this type can be null.
+  static const Annotation nullable =
+      // Any other valid `Nullable` annotation would work.
+      Annotation(binaryName: 'androidx.annotation.Nullable');
+
+  /// Specifies that this type cannot be null.
+  static const Annotation nonNull =
+      // Any other valid `NonNull` annotation would work.
+      Annotation(binaryName: 'androidx.annotation.NonNull');
+
+  const Annotation({
     required this.binaryName,
     this.properties = const {},
+    this.typePath = const [],
   });
 
   final String binaryName;
   final Map<String, Object> properties;
+
+  @JsonKey(fromJson: typePathFromString)
+  final List<TypePathStep> typePath;
 
   factory Annotation.fromJson(Map<String, dynamic> json) =>
       _$AnnotationFromJson(json);