+++ pkgs/jnigen/test/kotlin_test/runtime_test_registrant.dart
@@ -13,11 +13,11 @@ void registerTests(String groupName, TestRunnerCallback test) {
     test('Suspend functions', () async {
       await using((arena) async {
         final suspendFun = SuspendFun()..releasedBy(arena);
-        final hello = await suspendFun.sayHello();
+        final hello = (await suspendFun.sayHello())!;
         expect(hello.toDartString(releaseOriginal: true), 'Hello!');
         const name = 'Bob';
         final helloBob =
-            await suspendFun.sayHello$1(name.toJString()..releasedBy(arena));
+            (await suspendFun.sayHello$1(name.toJString()..releasedBy(arena)))!;
         expect(helloBob.toDartString(releaseOriginal: true), 'Hello $name!');
       });
     });